requirements.txt
pyaml
simplekml
requests-cache
tqdm
flake8
